f3baa00 Fix int32/64 problem in multisig payment proposal parsing (Jonas Schnelli)

Tree-SHA512: 96105cff740d8d769c89f6828cadddfa529bf5e407966ad98bbb1c432192c47d03081bec9bf23ca755d3aadd4546bc8a28f67df0cc986df95ebe9d6f76765b65

+++ src/bitpaywalletclient/bpwalletclient.h
@@ -110,7 +110,7 @@ public:
     bool GetFeeLevels();
 
     //!get the feeperkb rate for the given prio (0 = priority, 1 = normal, 2 = economy)
-    int GetFeeForPriority(int prio = 0);
+    int64_t GetFeeForPriority(int prio = 0);
 
     //!load available wallets over wallet server
     bool GetWallets(std::string& response);
